moved all ajax functions to /rest/
