+++ src/rest/password.php
@@ -1,14 +1,7 @@
 <?php
-header('Content-Type: application/json');
-function error($msg) {
-    echo json_encode(array("status" => "error", "message" => $msg));
-    die();
-}
-if ($_SERVER['REQUEST_METHOD'] !== 'POST'){
-    error('Wrong method');
-}
-require_once("function/sqllink.php");
-require_once("function/encryption.php");
+require_once(dirname(__FILE__).'/../function/ajax.php');
+require_once(dirname(__FILE__).'/../function/sqllink.php');
+require_once(dirname(__FILE__).'/../function/encryption.php');
 $link=sqllink();
 if(!checksession($link)) {
     error("AUTHENTICATION ERROR, PLEASE RELOGIN");
